# backend/specs/openapi.yaml
openapi: 3.0.3
info:
  title: 智能测试管理平台 API
  description: |
    智能测试管理平台 RESTful API 定义。

    ## 认证
    所有 API（除健康检查和认证接口外）需要 JWT Token 认证：
    ```
    Authorization: Bearer <token>
    ```

    ## 错误响应
    统一错误响应格式：
    ```json
    {
      "error": "错误信息"
    }
    ```
  version: 1.0.0
  contact:
    name: AITestOS Team
    email: support@aitestos.com
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.aitestos.com/api/v1
    description: 生产环境
...
paths:
...
  /plans/{id}:
    get:
      tags: [TestPlans]
      summary: 获取测试计划详情
      operationId: getTestPlan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanDetail'
        '404':
          description: 测试计划不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [TestPlans]
      summary: 更新测试计划
      operationId: updateTestPlan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTestPlanRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestPlan'
    patch:
      tags: [TestPlans]
      summary: 更新测试计划状态
      operationId: updatePlanStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [draft, active, completed, archived]
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestPlan'
        '400':
          description: 状态转换不合法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 计划不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [TestPlans]
      summary: 删除测试计划
      operationId: deleteTestPlan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 删除成功
...
  /plans/{id}/results:
    get:
      tags: [TestPlans]
      summary: 获取计划执行结果
      operationId: getPlanResults
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 测试计划 ID
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestResult'
    post:
      tags: [TestPlans]
      summary: 录入测试结果
      operationId: recordResult
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 测试计划 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordResultRequest'
      responses:
        '201':
          description: 录入成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResult'

  # ==================== AI 生成 ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CaseStatus:
      type: string
      enum: [unexecuted, pass, block, fail]
    CaseType:
      type: string
      enum: [functionality, performance, api, ui, security]
    PlanStatus:
      type: string
      enum: [draft, active, completed, archived]
    Priority:
      type: string
      enum: [P0, P1, P2, P3]
    ResultStatus:
      type: string
      enum: [pass, fail, block, skip]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 错误信息

    # ==================== 认证 ====================
...
    TestCase:
      type: object
      properties:
        id:
          type: string
          format: uuid
        moduleId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
        preconditions:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            type: string
        expected:
          type: object
        caseType:
          $ref: '#/components/schemas/CaseType'
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/CaseStatus'
        aiMetadata:
          type: object
          description: AI 生成元数据
          properties:
            generationTaskId:
              type: string
              format: uuid
            confidence:
              type: string
              enum: [high, medium, low]
            referencedChunks:
              type: array
              items:
                type: object
                properties:
                  chunkId:
                    type: string
                    format: uuid
                  documentId:
                    type: string
                    format: uuid
                  documentTitle:
                    type: string
                  similarityScore:
                    type: number
                    format: double
            modelVersion:
              type: string
            generatedAt:
              type: string
              format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    TestPlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/PlanStatus'
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PlanDetail:
      type: object
      description: 测试计划详情（含用例、结果和统计）
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/PlanStatus'
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        cases:
          type: array
          items:
            $ref: '#/components/schemas/TestCase'
        results:
          type: array
          items:
            $ref: '#/components/schemas/TestResult'
        stats:
          $ref: '#/components/schemas/PlanStatistics'
    PlanStatistics:
      type: object
      properties:
        total_cases:
          type: integer
        passed_cases:
          type: integer
        failed_cases:
          type: integer
        blocked_cases:
          type: integer
        skipped_cases:
          type: integer
        unexecuted:
          type: integer
...
    UpdateTestPlanRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 255
        description:
          type: string

    # ==================== 测试结果 ====================
    TestResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        planId:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        executedBy:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ResultStatus'
        note:
          type: string
        executedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RecordResultRequest:
      type: object
      required: [plan_id, case_id, status]
      properties:
        plan_id:
          type: string
          format: uuid
        case_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pass, fail, block, skip]
        note:
          type: string

    # ==================== 生成任务 ====================
...